Split Host and Component management
* Component management is on its own page now
* Components per host is still viewable on the host page
* Add ability to update host and component
* Add auth to ensure calls that require admin are protected

+++ service/src/main/java/org/kiwiproject/champagne/resource/HostConfigurationResource.java
@@ -1,9 +1,5 @@
 package org.kiwiproject.champagne.resource;
 
-import static java.util.Objects.isNull;
-import static org.apache.commons.lang3.StringUtils.isBlank;
-import static org.kiwiproject.champagne.util.DeployableSystems.getSystemIdOrThrowBadRequest;
-
 import com.codahale.metrics.annotation.ExceptionMetered;
 import com.codahale.metrics.annotation.Timed;
 import org.apache.commons.lang3.StringUtils;
@@ -16,18 +12,16 @@ import org.kiwiproject.champagne.model.Host;
 import org.kiwiproject.dropwizard.error.dao.ApplicationErrorDao;
 import org.kiwiproject.jaxrs.exception.JaxrsNotFoundException;
 
-import java.util.List;
 import javax.annotation.security.PermitAll;
-import javax.ws.rs.Consumes;
-import javax.ws.rs.DELETE;
-import javax.ws.rs.GET;
-import javax.ws.rs.POST;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
-import javax.ws.rs.QueryParam;
+import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
+import java.util.List;
+
+import static java.util.Objects.isNull;
+import static org.apache.commons.lang3.StringUtils.isBlank;
+import static org.kiwiproject.champagne.util.DeployableSystems.checkUserAdminOfSystem;
+import static org.kiwiproject.champagne.util.DeployableSystems.getSystemIdOrThrowBadRequest;
 
 @Path("/host")
 @Consumes(MediaType.APPLICATION_JSON)
@@ -60,6 +54,8 @@ public class HostConfigurationResource extends AuditableResource {
     @Timed
     @ExceptionMetered
     public Response createHost(Host host) {
+        checkUserAdminOfSystem();
+
         if (isNull(host.getDeployableSystemId())) {
             var systemId = getSystemIdOrThrowBadRequest();
             host = host.withDeployableSystemId(systemId);
@@ -72,11 +68,29 @@ public class HostConfigurationResource extends AuditableResource {
         return Response.accepted().build();
     }
 
+    @PUT
+    @Path("/{id}")
+    @Timed
+    @ExceptionMetered
+    public Response updateHost(Host host, @PathParam("id") Long hostId) {
+        checkUserAdminOfSystem();
+
+        var updateCount = hostDao.updateHost(host.getHostname(), StringUtils.join(host.getTags(), ","), hostId);
+
+        if (updateCount > 0) {
+            auditAction(hostId, Host.class, Action.UPDATED);
+        }
+
+        return Response.accepted().build();
+    }
+
     @DELETE
     @Path("/{id}")
     @Timed
     @ExceptionMetered
     public Response deleteHost(@PathParam("id") Long id) {
+        checkUserAdminOfSystem();
+
         var deleteCount = hostDao.deleteHost(id);
 
         if (deleteCount > 0) {
@@ -84,6 +98,17 @@ public class HostConfigurationResource extends AuditableResource {
         }
 
         return Response.accepted().build();
+    }
+
+    @GET
+    @Path("/components")
+    @Timed
+    @ExceptionMetered
+    public Response listComponents() {
+        var systemId = getSystemIdOrThrowBadRequest();
+        var components = componentDao.findComponentsForSystem(systemId);
+
+        return Response.ok(components).build();
     }
 
     @GET
@@ -102,6 +127,8 @@ public class HostConfigurationResource extends AuditableResource {
     @Timed
     @ExceptionMetered
     public Response createComponent(Component component) {
+        checkUserAdminOfSystem();
+
         if (isNull(component.getDeployableSystemId())) {
             var systemId = getSystemIdOrThrowBadRequest();
             component = component.withDeployableSystemId(systemId);
@@ -114,11 +141,29 @@ public class HostConfigurationResource extends AuditableResource {
         return Response.accepted().build();
     }
 
+    @PUT
+    @Path("/component/{id}")
+    @Timed
+    @ExceptionMetered
+    public Response updateComponent(Component component, @PathParam("id") Long componentId) {
+        checkUserAdminOfSystem();
+
+        var updateCount = componentDao.updateComponent(component.getComponentName(), component.getTag(), componentId);
+
+        if (updateCount > 0) {
+            auditAction(componentId, Component.class, Action.UPDATED);
+        }
+
+        return Response.accepted().build();
+    }
+
     @DELETE
     @Path("/component/{componentId}")
     @Timed
     @ExceptionMetered
     public Response deleteComponent(@PathParam("componentId") Long componentId) {
+        checkUserAdminOfSystem();
+
         var deleteCount = componentDao.deleteComponent(componentId);
 
         if (deleteCount > 0) {
